backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
googletrans==4.0.0-rc1
requests==2.31.0
ollama==0.1.6
python-multipart==0.0.6
textract==1.6.5